refactor: GetEnrollmentResponse 도메인 위치 정리
